[fix] move files in ErrorPage
