+++ src/container/ErrorPage/components/IconError/data/Animation404.json
