+++ src/container/ErrorPage/components/IconError/index.style.ts
