+++ src/container/ErrorPage/components/Info/index.style.ts
